re-org into modules and optimize scoring
