--- data/refseq_chr_map.txt
@@ -1,25 +1,0 @@
-1	chr1	NC_000001.11
-2	chr2	NC_000002.12
-3	chr3	NC_000003.12
-4	chr4	NC_000004.12
-5	chr5	NC_000005.10
-6	chr6	NC_000006.12
-7	chr7	NC_000007.14
-8	chr8	NC_000008.11
-9	chr9	NC_000009.12
-10	chr10	NC_000010.11
-11	chr11	NC_000011.10
-12	chr12	NC_000012.12
-13	chr13	NC_000013.11
-14	chr14	NC_000014.9
-15	chr15	NC_000015.10
-16	chr16	NC_000016.10
-17	chr17	NC_000017.11
-18	chr18	NC_000018.10
-19	chr19	NC_000019.10
-20	chr20	NC_000020.11
-21	chr21	NC_000021.9
-22	chr22	NC_000022.11
-X	chrX	NC_000023.11
-Y	chrY	NC_000024.10
-M	chrM	NC_012920.1
